+++ src/layout/Input/InputComponent.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { numericFormatter, patternFormatter, removeNumericFormat, removePatternFormat } from 'react-number-format';
+import React, { useCallback, useState } from 'react';
 import type { NumericFormatProps, PatternFormatProps } from 'react-number-format';
 
 import { SearchField } from '@altinn/altinn-design-system';
@@ -9,49 +8,12 @@ import { useDataModelBindings } from 'src/features/formData/useDataModelBindings
 import { useLanguage } from 'src/features/language/useLanguage';
 import { useMapToReactNumberConfig } from 'src/hooks/useMapToReactNumberConfig';
 import { useRerender } from 'src/hooks/useReload';
+import { getCleanValue, getFormattedValue, isEmptyObject } from 'src/layout/Input/number-format-helpers';
 import { useCharacterLimit } from 'src/utils/inputUtils';
 import type { PropsFromGenericComponent } from 'src/layout';
-import type { IInputFormatting, NumberFormatProps } from 'src/layout/Input/config.generated';
+import type { IInputFormatting } from 'src/layout/Input/config.generated';
 
 export type IInputProps = PropsFromGenericComponent<'Input'>;
-
-function isEmptyObject(obj: Record<string, any>): boolean {
-  return Object.keys(obj).length === 0;
-}
-
-const getChangeMeta = (newValue: string) => ({
-  from: { start: 0, end: 0 },
-  to: { start: 0, end: newValue.length },
-  lastValue: '',
-});
-
-const getCleanValue = (newValue: string, inputFormatting: IInputFormatting) => {
-  if (isNumericFormat(inputFormatting)) {
-    return removeNumericFormat(newValue, getChangeMeta(newValue), {
-      ...inputFormatting,
-    });
-  }
-  return removePatternFormat(newValue, getChangeMeta(newValue), {
-    ...(inputFormatting as PatternFormatProps),
-  });
-};
-
-const getFormattedValue = (newValue: string, inputFormatting: IInputFormatting): string => {
-  const cleanedValue = getCleanValue(newValue, inputFormatting);
-
-  if (isNumericFormat(inputFormatting) && inputFormatting.number) {
-    if ((inputFormatting.number as NumberFormatProps).allowedDecimalSeparators) {
-      return numericFormatter(cleanedValue, inputFormatting);
-    }
-
-    return numericFormatter(cleanedValue, inputFormatting.number);
-  }
-
-  if (isPatternFormat(inputFormatting) && !isEmptyObject(inputFormatting)) {
-    return patternFormatter(cleanedValue, inputFormatting);
-  }
-  return newValue;
-};
 
 export const isPatternFormat = (
   numberFormat: NumericFormatProps | PatternFormatProps,
@@ -80,8 +42,10 @@ export function InputComponent({ node, isValid, overrideDisplay }: IInputProps) 
     debounce,
   } = useDataModelBindings(dataModelBindings, saveWhileTyping);
 
-  const [localValue, setLocalValue] = useState<string>();
   const reactNumberFormatConfig = useMapToReactNumberConfig(formatting as IInputFormatting | undefined, formValue);
+
+  const [localValue, setLocalValue] = useState<string>(getFormattedValue(formValue, reactNumberFormatConfig));
+
   const [inputKey, rerenderInput] = useRerender('input');
 
   const onBlur = useCallback(() => {
@@ -93,17 +57,14 @@ export function InputComponent({ node, isValid, overrideDisplay }: IInputProps) 
 
   const ariaLabel = overrideDisplay?.renderedInTable === true ? langAsString(textResourceBindings?.title) : undefined;
 
-  useEffect(() => {
-    valueChanged(formValue);
-  });
   const valueChanged = (newValue: string) => {
-    if (!formatting) {
+    if (isEmptyObject(reactNumberFormatConfig)) {
       setLocalValue(newValue);
       setValue('simpleBinding', newValue);
       return;
     }
-    setLocalValue(getFormattedValue(newValue, formatting));
-    setValue('simpleBinding', getCleanValue(newValue, formatting));
+    setLocalValue(getFormattedValue(newValue, reactNumberFormatConfig));
+    setValue('simpleBinding', getCleanValue(newValue, reactNumberFormatConfig));
   };
 
   if (variant === 'search') {
